symptom_predictor_ml/requirements.txt
# Core scientific stack
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1

# Machine learning
scikit-learn==1.4.2
xgboost==2.0.3
joblib==1.3.2
threadpoolctl==3.5.0

# Data handling
python-dateutil==2.9.0.post0
pytz==2024.1
tzdata==2024.1
six==1.16.0

# API (if you ever serve via FastAPI / Uvicorn)
fastapi==0.110.0
uvicorn==0.27.1
starlette==0.36.3
anyio==4.3.0
idna==3.7
sniffio==1.3.1
colorama==0.4.6
click==8.1.7

# Optional: MySQL connector if needed
mysql-connector-python==9.0.0
